requirements.txt
requests
sendgrid
pyyaml